First pass at Bindl::Encrypt.
Right now, it works when all goes well.

It doesn't capture what GPG prints to stderr, or its exit code, so
errors just cause huge issues. I think the solution is in
Open3.popen3.

I added a few helper methods to Bindl::Entry in the form of #get
and #set, defined in their respective modules. These are helpful
when playing in a REPL.